+++ src/Data/Bytes/IO.hs
@@ -7,11 +7,13 @@
 
 module Data.Bytes.IO
   ( hGet
+  , hPut
   ) where
 
-import Data.Primitive (MutableByteArray)
+import Data.Primitive (MutableByteArray,ByteArray(..))
 import Data.Word (Word8)
 import Data.Bytes.Types (Bytes(Bytes))
+import Data.Bytes.Pure (pin,contents)
 import System.IO (Handle)
 import Foreign.Ptr (Ptr)
 import GHC.IO (IO(IO))
@@ -23,6 +25,14 @@ import qualified Data.Primitive as PM
 -- 'Bytes' are pinned. This is implemented with 'hGetBuf'.
 hGet :: Handle -> Int -> IO Bytes
 hGet h i = createPinnedAndTrim i (\p -> IO.hGetBuf h p i)
+
+-- | Outputs 'Bytes' to the specified 'Handle'. This is implemented
+-- with 'hPutBuf'.
+hPut :: Handle -> Bytes -> IO ()
+hPut h b0 = do
+  let b1@(Bytes arr _ len) = pin b0
+  IO.hPutBuf h (contents b1) len
+  touchByteArrayIO arr
 
 -- Only used internally.
 createPinnedAndTrim :: Int -> (Ptr Word8 -> IO Int) -> IO Bytes
@@ -37,5 +47,9 @@ createPinnedAndTrim maxSz f = do
 
 touchMutableByteArrayIO :: MutableByteArray s -> IO ()
 touchMutableByteArrayIO (PM.MutableByteArray x) =
+  IO (\s -> (# Exts.touch# x s, () #))
+
+touchByteArrayIO :: ByteArray -> IO ()
+touchByteArrayIO (ByteArray x) =
   IO (\s -> (# Exts.touch# x s, () #))